uppgift2/website/requirements.txt
Flask==2.0.1
psycopg2-binary
